requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docopt==0.6.2
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
yarg==0.1.9
